Refactor ID types from uint32 to uint64 across the codebase
- Updated FileHeader, NodeHeader, and HNSWMetadata structures to use uint64 for IDs and counts.
- Modified functions in persistence, search, vectorstore, and WAL modules to handle uint64 IDs.
- Adjusted tests and benchmarks to reflect the changes in ID types.
- Ensured proper alignment checks for uint64 slices in safety.go.
- Updated serialization and deserialization methods to accommodate the new ID type.

+++ benchmark_test/isolated/hnsw_isolated_test.go
@@ -28,7 +28,7 @@ func BenchmarkHNSWIsolatedInsert(b *testing.B) {
 			vec := rng.UnitVector(dim)
 			// HNSW requires sequential IDs usually, or we manage them.
 			// HNSW.Insert takes (ctx, id, vector)
-			if err := idx.ApplyInsert(ctx, uint32(i), vec); err != nil {
+			if err := idx.ApplyInsert(ctx, uint64(i), vec); err != nil {
 				b.Fatal(err)
 			}
 		}
@@ -56,7 +56,7 @@ func BenchmarkHNSWIsolatedSearch(b *testing.B) {
 	for i := 0; i < size; i++ {
 		vec := rng.UnitVector(dim)
 		vectors[i] = vec
-		if err := idx.ApplyInsert(ctx, uint32(i), vec); err != nil {
+		if err := idx.ApplyInsert(ctx, uint64(i), vec); err != nil {
 			b.Fatal(err)
 		}
 	}
