+++ engine/map_store_test.go
@@ -25,7 +25,7 @@ func TestMapStore(t *testing.T) {
 	}
 
 	// Test BatchSet
-	batch := map[uint32]string{
+	batch := map[uint64]string{
 		2: "two",
 		3: "three",
 		4: "four",
@@ -39,7 +39,7 @@ func TestMapStore(t *testing.T) {
 	}
 
 	// Test BatchGet
-	results, err := st.BatchGet([]uint32{1, 2, 3, 999})
+	results, err := st.BatchGet([]uint64{1, 2, 3, 999})
 	if err != nil {
 		t.Fatalf("BatchGet failed: %v", err)
 	}
@@ -68,7 +68,7 @@ func TestMapStore(t *testing.T) {
 	}
 
 	// Test BatchDelete
-	if err := st.BatchDelete([]uint32{2, 3}); err != nil {
+	if err := st.BatchDelete([]uint64{2, 3}); err != nil {
 		t.Fatalf("BatchDelete failed: %v", err)
 	}
 
@@ -105,7 +105,7 @@ func TestMapStoreConcurrent(t *testing.T) {
 		wg.Add(1)
 		go func(id int) {
 			defer wg.Done()
-			_ = st.Set(uint32(id), id*2)
+			_ = st.Set(uint64(id), id*2)
 		}(i)
 	}
 
@@ -114,7 +114,7 @@ func TestMapStoreConcurrent(t *testing.T) {
 		wg.Add(1)
 		go func(id int) {
 			defer wg.Done()
-			_, _ = st.Get(uint32(id))
+			_, _ = st.Get(uint64(id))
 		}(i)
 	}
