+++ engine/store.go
@@ -9,26 +9,24 @@ package engine
 type Store[T any] interface {
 	// Get retrieves the data associated with the given ID.
 	// Returns the data and true if found, or zero value and false if not found.
-	Get(id uint32) (T, bool)
+	Get(id uint64) (T, bool)
 
 	// Set stores data associated with the given ID.
 	// If the ID already exists, it updates the data.
-	Set(id uint32, data T) error
-
+	Set(id uint64, data T) error
 	// Delete removes the data associated with the given ID.
 	// Returns an error if the ID doesn't exist.
-	Delete(id uint32) error
+	Delete(id uint64) error
 
 	// BatchGet retrieves data for multiple IDs in a single operation.
 	// Returns a map of id -> data for all found IDs.
-	BatchGet(ids []uint32) (map[uint32]T, error)
-
+	BatchGet(ids []uint64) (map[uint64]T, error)
 	// BatchSet stores multiple id -> data pairs in a single operation.
 	// If any operation fails, the entire batch may be rolled back (implementation-dependent).
-	BatchSet(items map[uint32]T) error
+	BatchSet(items map[uint64]T) error
 
 	// BatchDelete removes data for multiple IDs in a single operation.
-	BatchDelete(ids []uint32) error
+	BatchDelete(ids []uint64) error
 
 	// Len returns the number of items currently stored.
 	Len() int
@@ -37,5 +35,5 @@ type Store[T any] interface {
 	Clear() error
 
 	// ToMap returns a copy of all data as a map (for serialization).
-	ToMap() map[uint32]T
+	ToMap() map[uint64]T
 }
